api release

+++ api.py
@@ -5,28 +5,28 @@ import xgboost as xgb
 import joblib
 
 class DiabetesInput(BaseModel):
-    Diabetes_012: int
-    HighBP: int
-    HighChol: int
-    CholCheck: int
+    Diabetes_012: float
+    HighBP: float
+    HighChol: float
+    CholCheck: float
     BMI: float
-    Smoker: int
-    Stroke: int
-    HeartDiseaseorAttack: int
-    PhysActivity: int
-    Fruits: int
-    Veggies: int
-    HvyAlcoholConsump: int
-    AnyHealthcare: int
-    NoDocbcCost: int
-    GenHlth: int
-    MentHlth: int
-    PhysHlth: int
-    DiffWalk: int
-    Sex: int
-    Age: int
-    Education: int
-    Income: int
+    Smoker: float
+    Stroke: float
+    HeartDiseaseorAttack: float
+    PhysActivity: float
+    Fruits: float
+    Veggies: float
+    HvyAlcoholConsump: float
+    AnyHealthcare: float
+    NoDocbcCost: float
+    GenHlth: float
+    MentHlth: float
+    PhysHlth: float
+    DiffWalk: float
+    Sex: float
+    Age: float
+    Education: float
+    Income: float
 
 app = FastAPI()
 model = joblib.load("diabetes_model.joblib")
